requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==45.0.5
easygui==0.98.3
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.54.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
lark==1.1.9
lxml==6.0.0
msoffcrypto-tool==5.4.2
olefile==0.47
oletools==0.60.2
openpyxl==3.1.5
pcodedmp==1.2.6
pdfminer.six==20250506
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-docx==1.2.0
python-multipart==0.0.20
red-black-tree-mod==1.22
RTFDE==0.1.2.1
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
striprtf==0.0.29
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
win_unicode_console==0.5
xlrd==2.0.2
